[FEATURE] Create Todo Items (#3)
* [FEATURE] Create TodoItem factory

Sets up a factory to create TodoItems and an acceptance test
in the root index.js to ensure the proper object is returned.

Removes main.js from the distribution folder. This clogs up the
commit history pretty badly.

* [FEATURE] Expand acceptance test to multiple TodoItems

In order to have a bigger base to test with, extended the dummy
todoItems to a count of 3 and used `console.dir` to more quickly
display the collection of TodoItem objects when visiting the
browser console.